chore: linting savepoint
