+++ Makefile
@@ -23,6 +23,7 @@ check: ## Run code quality tools.
 	@poetry run mypy
 	@echo "🚀 Checking for obsolete dependencies: Running deptry"
 	@poetry run deptry .
+	@make spec
 
 .PHONY: clean
 clean:
@@ -201,8 +202,8 @@ pingpg:
 pingdb:
 	@psql "postgresql://alexanderpatrie:dev@localhost:65432/sms?sslmode=disable"
 
-.PHONY: check-latest 
+.PHONY: check-latest
 check-latest:
 	@curl -s https://api.github.com/repos/CovertLab/vEcoli/commits/master | jq -r '"\(.sha[0:7]) \(.commit.author.date)"'
-	
+
 .DEFAULT_GOAL := help
